Rename framework, dropped the ‘iOS’
