+++ EvernoteSDK/Info.plist
