+++ EvernoteSDKSample/EvernoteSDKSample/SaveActivityViewController.m
@@ -7,7 +7,7 @@
 //
 
 #import "SaveActivityViewController.h"
-#import <EvernoteSDKiOS/EvernoteSDKiOS.h>
+#import <EvernoteSDK/EvernoteSDK.h>
 
 @interface SaveActivityViewController () <ENSaveToEvernoteActivityDelegate>
