+++ EvernoteSDKSample/EvernoteSDKSample/UserInfoViewController.m
@@ -7,7 +7,7 @@
 //
 
 #import "UserInfoViewController.h"
-#import <EvernoteSDKiOS/EvernoteSDKiOS.h>
+#import <EvernoteSDK/EvernoteSDK.h>
 
 @interface UserInfoViewController ()
