+++ evernote-sdk-ios/ENSDK/SendToEvernoteActivity/ENSaveToEvernoteViewController.h
@@ -27,7 +27,7 @@
  */
 
 #import <UIKit/UIKit.h>
-#import "EvernoteSDKiOS.h"
+#import "EvernoteSDK.h"
 @class ENSaveToEvernoteViewController;
 
 @protocol ENSendToEvernoteViewControllerDelegate <NSObject>
