+++ evernote-sdk-ios/ENSDK/SendToEvernoteActivity/ENSaveToEvernoteViewController.m
@@ -29,7 +29,7 @@
 #import "ENSaveToEvernoteViewController.h"
 #import "ENNotebookChooserViewController.h"
 #import "ENNotebookPickerButton.h"
-#import "EvernoteSDKiOS.h"
+#import "EvernoteSDK.h"
 #import "ENTheme.h"
 #import "RMSTokenView.h"
 #import "ENNotebookPickerView.h"
